requirements.txt

streamlit
sentence-transformers
python-docx
python-pptx
PyPDF2
PyMuPDF
numpy
pickle-mixin
